+++ automationTesting-Firefox.xml
@@ -3,7 +3,7 @@
 <suite name="FirefoxAutomationTesting">
 	<test name="All_Tests">
 		<packages>
-			<package name="tests" />
+			<package name="tests.*" />
 		</packages>
 	</test>
 </suite>
