+++ automationTesting-IE.xml
@@ -3,7 +3,7 @@
 <suite name="IEAutomationTesting">
 	<test name="All_Tests">
 		<packages>
-			<package name="tests" />
+			<package name="tests.*" />
 		</packages>
 	</test>
 </suite>
