+++ src/tests/homePage/LogoSearchCart.java
@@ -1,19 +1,20 @@
-package tests;
+package tests.homePage;
 
-import org.junit.Assert;
+
 import org.openqa.selenium.WebElement;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import pageObjects.HomePage;
+import pageObjects.homePage.LogoSearchCartPage;
 import testUtilities.TestBase;
 
-public class Home extends TestBase {
+public class LogoSearchCart extends TestBase {
 	@Test(description = "Test that the logo exists and contains the correct image")
 	public void findLogo() {
 		//** 
 		//** Arrange
 		//**
-		HomePage homePage = new HomePage(this.driver);
+		LogoSearchCartPage homePage = new LogoSearchCartPage(this.driver);
 		
 		//** 
 		//** Act
@@ -23,8 +24,8 @@ public class Home extends TestBase {
 		//** 
 		//** Assert
 		//**
-		Assert.assertNotNull("Logo exists", logo);
-		Assert.assertTrue("Logo contains the correct image", logo.getAttribute("src").contains("logo.jpg"));
+		Assert.assertNotNull(logo, "Logo exists");
+		Assert.assertTrue(logo.getAttribute("src").contains("logo.jpg"), "Logo contains the correct image");
 	}
 	
 	@Test(description = "Test that the search textbox exists")
@@ -32,7 +33,7 @@ public class Home extends TestBase {
 		//** 
 		//** Arrange
 		//**
-		HomePage homePage = new HomePage(this.driver);
+		LogoSearchCartPage homePage = new LogoSearchCartPage(this.driver);
 		
 		//** 
 		//** Act
@@ -42,6 +43,6 @@ public class Home extends TestBase {
 		//** 
 		//** Assert
 		//**
-		Assert.assertNotNull("Search textbox exists", searchTextBox);
+		Assert.assertNotNull(searchTextBox, "Search textbox exists");
 	}
 }
