+++ src/tests/login/Login.java
@@ -1,12 +1,12 @@
-package tests;
+package tests.login;
 
-import org.junit.Assert;
 import org.openqa.selenium.WebElement;
+import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import pageObjects.HomePage;
-import pageObjects.LoginPage;
-import pageObjects.MyAccountPage;
+import pageObjects.homePage.LogoSearchCartPage;
+import pageObjects.loginPage.LoginPage;
+import pageObjects.myAccountPage.MyAccountPage;
 import testUtilities.TestBase;
 
 public class Login extends TestBase {
@@ -20,7 +20,7 @@ public class Login extends TestBase {
 		String password = getProperties().getProperty("defaultPassword");
 		
 		// open home page
-		HomePage homePage = new HomePage(this.getDriver());
+		LogoSearchCartPage homePage = new LogoSearchCartPage(this.getDriver());
 		
 		// click the login link and get sent to the login page
 		LoginPage loginPage = homePage.clickLogin();
@@ -35,6 +35,6 @@ public class Login extends TestBase {
 		//**Assert
 		//**
 		WebElement link = myAccountPage.getSignOutLink();
-		Assert.assertNotNull("The sign out link now appears", link);
+		Assert.assertNotNull(link, "The sign out link now appears");
 	}
 }
